adding play and stop music for menu

+++ stage.cpp
@@ -106,17 +106,17 @@ void resetStage(void)
 
 }
 
-void calcSlope(int x1, int y1, int x2, int y2, float* dx, float* dy)
+void calcSlope(int x1, int y1, int x2, int y2, double* dx, double* dy)
 {
-    int steps = max(abs(x1 - x2), abs(y1 - y2));
+    double steps = max(abs(x1 - x2), abs(y1 - y2));
     if (steps == 0)
     {
         *dx = *dy = 0;
         return;
     }
 
-    *dx = (x1 - x2) / (float)steps;
-    *dy = (y1 - y2) / (float)steps;
+    *dx = (x1 - x2) / steps;
+    *dy = (y1 - y2) / steps;
 }
 
 void clipPlayer()
